refactor(agent): split into an agent app and an engine-agnostic harness runtime
Address the god-module and the misleading package name:

- services/oss/src/harness/ (was agent_pi/): the engine-agnostic runtime — ports.py,
  transports.py (was harness.py), environment.py, wire.py. Named for the seam, not Pi;
  harness choice (pi/claude) lives inside the runtime, so there is no agent_claude.
- services/oss/src/agent/ (was the 470-line agent.py god-module): the Agenta workflow app
  — app.py (thin handler + backend wiring), inputs.py (request parsing), tools.py,
  secrets.py, tracing.py, client.py (shared backend access), schemas.py, config.py.

No behavior change. Verified live: a playground run answers 'REFACTOR-OK Lisbon' with usage.
