+++ services/oss/src/agent/__init__.py
@@ -1,0 +1,12 @@
+"""The Agenta agent workflow app and its glue.
+
+The handler and harness wiring are in ``app``; request parsing in ``inputs``; tool
+resolution in ``tools``; provider secrets in ``secrets``; trace/usage glue in ``tracing``;
+the ``/inspect`` schemas in ``schemas``; the file-backed defaults in ``config``. The
+engine-agnostic runtime (the harness/environment seams and adapters) lives in
+``oss.src.harness``.
+"""
+
+from oss.src.agent.app import agent_v0_app, create_agent_app
+
+__all__ = ["agent_v0_app", "create_agent_app"]
