+++ services/oss/src/harness/__init__.py
@@ -1,18 +1,18 @@
-"""Agent runtime: ports and adapters for the agent service.
+"""Engine-agnostic agent runtime: the harness and environment seams, plus their adapters.
 
-The Python service is "our agent implementation". It owns two seams (see
+Nothing here is Agenta-specific. The Agenta workflow integration (the ``/invoke`` handler,
+tool resolution, secrets, tracing) lives in ``oss.src.agent``. Two seams (see
 ``docs/design/agent-workflows/harness-port-redesign/``):
 
-- ``Harness``: the agent engine. ``SubprocessHarness`` and ``HttpHarness`` (in
-  ``harness.py``) are the two transports to the TypeScript runner; the engine (legacy
-  in-process Pi vs rivet over ACP) is an env value, not a class. ``create_session``
-  returns an :class:`AgentSession`, the rivet-shaped session abstraction.
+- ``Harness``: the agent engine. ``SubprocessHarness`` and ``HttpHarness`` (``transports.py``)
+  reach the TypeScript runner over a subprocess or HTTP. The engine that runs behind them
+  (rivet over ACP, or the legacy in-process Pi path) is an env value, not a class.
+  ``create_session`` returns an :class:`AgentSession` (create / prompt / destroy).
 - ``Environment``: where the harness process runs. ``LocalEnvironment`` runs it as a local
   subprocess; a sandbox environment is selected inside the rivet runner.
 """
 
 from .environment import LocalEnvironment
-from .harness import HttpHarness, SubprocessHarness
 from .ports import (
     AgentEvent,
     AgentRequest,
@@ -20,13 +20,14 @@ from .ports import (
     AgentSession,
     ContentBlock,
     Environment,
-    HarnessCapabilities,
     Harness,
+    HarnessCapabilities,
     Message,
     SessionConfig,
     ToolCallback,
     TraceContext,
 )
+from .transports import HttpHarness, SubprocessHarness
 
 __all__ = [
     "AgentEvent",
